Refactoring to be more simple and less error prone. Split up functions into multiple files.
